Move integration tests per setup logic and extract utils.
