+++ test/integration-test.ts
@@ -1,178 +1,27 @@
 import * as cp from 'child_process';
 import * as path from 'path';
 import * as fs from 'fs';
-import { createTempDir } from 'broccoli-test-helper';
-import { URI } from 'vscode-uri';
 
-import { createMessageConnection, MessageConnection, Logger, StreamMessageReader, StreamMessageWriter } from 'vscode-jsonrpc';
 import {
-  DidOpenTextDocumentNotification,
-  InitializeRequest,
-  CompletionRequest,
-  DefinitionRequest,
-  DocumentSymbolRequest,
-  ExecuteCommandRequest,
-  Definition,
-  ReferencesRequest,
-} from 'vscode-languageserver-protocol';
+  startServer,
+  initServer,
+  getResult,
+  reloadProjects,
+  makeAddonPackage,
+  makeProject,
+  createProject,
+  normalizeRegistry,
+  normalizePath,
+} from './test_helpers/integration-helpers';
+import { createMessageConnection, MessageConnection, Logger, StreamMessageReader, StreamMessageWriter } from 'vscode-jsonrpc';
 
+import { CompletionRequest, DefinitionRequest, DocumentSymbolRequest, ExecuteCommandRequest, ReferencesRequest } from 'vscode-languageserver-protocol';
 type UnknownResult = Record<string, unknown>;
 type Registry = {
   [key: string]: {
     [key: string]: string[];
   };
 };
-
-function startServer() {
-  return cp.spawn('node_modules/.bin/nyc', ['--reporter', 'none', 'node', './inst/start-server.js', '--stdio'], {
-    cwd: path.join(__dirname, '..'),
-  });
-}
-
-async function reloadProjects(connection, project = undefined) {
-  const result = await connection.sendRequest(ExecuteCommandRequest.type, {
-    command: 'els.reloadProject',
-    arguments: project ? [project] : [],
-  });
-
-  return result;
-}
-
-async function createProject(files, connection): Promise<{ normalizedPath: string; result: UnknownResult; destroy(): void }> {
-  const dir = await createTempDir();
-
-  dir.write(files);
-  const normalizedPath = path.normalize(dir.path());
-  const result = (await connection.sendRequest(ExecuteCommandRequest.type, ['els:registerProjectPath', normalizedPath])) as {
-    registry: Registry;
-  };
-
-  return {
-    normalizedPath,
-    result,
-    destroy: async () => {
-      await dir.dispose();
-    },
-  };
-}
-
-function textDocument(modelPath, position = { line: 0, character: 0 }) {
-  const params = {
-    textDocument: {
-      uri: URI.file(modelPath).toString(),
-    },
-    position,
-  };
-
-  return params;
-}
-
-async function getResult(reqType, connection, files, fileToInspect, position) {
-  const { normalizedPath, destroy, result } = await createProject(files, connection);
-  const modelPath = path.join(normalizedPath, fileToInspect);
-  const params = textDocument(modelPath, position);
-
-  openFile(connection, modelPath);
-  const response = await connection.sendRequest(reqType, params);
-
-  await destroy();
-
-  return { response: normalizeUri(response, normalizedPath), registry: normalizeRegistry(normalizedPath, result.registry as Registry) };
-}
-
-function openFile(connection: MessageConnection, filePath: string) {
-  connection.sendNotification(DidOpenTextDocumentNotification.type, {
-    textDocument: {
-      uri: URI.file(filePath).toString(),
-      text: fs.readFileSync(filePath, 'utf8'),
-    },
-  });
-}
-
-function normalizePath(file: string) {
-  return file.split('\\').join('/');
-}
-
-function replaceTempUriPart(uri: string, base: string) {
-  const fsPath = normalizePath(URI.parse(uri).fsPath);
-  const basePath = normalizePath(URI.parse(base).fsPath);
-
-  return fsPath.split(basePath).pop();
-}
-
-function normalizeRegistry(root: string, registry: Registry) {
-  const normalizedRegistry: Registry = {};
-
-  Object.keys(registry).forEach((key) => {
-    normalizedRegistry[key] = {};
-    Object.keys(registry[key]).forEach((name) => {
-      normalizedRegistry[key][name] = registry[key][name].map((el) => normalizePath(path.relative(root, el)));
-    });
-
-    if (!Object.keys(normalizedRegistry[key]).length) {
-      delete normalizedRegistry[key];
-    }
-  });
-
-  return normalizedRegistry;
-}
-
-function normalizeUri(objects: Definition, base?: string) {
-  if (objects === null) {
-    return objects;
-  }
-
-  if (!Array.isArray(objects)) {
-    if (objects.uri) {
-      // objects.uri = replaceDynamicUriPart(objects.uri);
-
-      if (base) {
-        objects.uri = replaceTempUriPart(objects.uri, base);
-      }
-    }
-
-    return objects;
-  }
-
-  return objects.map((object) => {
-    if (object === null) {
-      return object;
-    }
-
-    return normalizeUri(object, base);
-  });
-}
-
-function makeProject(appFiles = {}, addons = {}) {
-  const dependencies = {};
-  const node_modules = {};
-
-  Object.keys(addons).forEach((name) => {
-    dependencies[name] = '*';
-    node_modules[name] = addons[name];
-  });
-  const fileStructure = Object.assign({}, appFiles, {
-    node_modules,
-    'package.json': JSON.stringify({
-      dependencies,
-    }),
-  });
-
-  return fileStructure;
-}
-
-function makeAddonPackage(name, config, addonConfig = undefined) {
-  const pack = {
-    name,
-    'ember-language-server': config,
-  };
-
-  if (addonConfig) {
-    pack['ember-addon'] = addonConfig;
-  }
-
-  return JSON.stringify(pack);
-}
 
 describe('integration', function () {
   let connection: MessageConnection;
@@ -205,201 +54,8 @@ describe('integration', function () {
 
   describe('Initialize request', () => {
     it('returns an initialize request', async () => {
-      const params = {
-        rootUri: URI.file(path.join(__dirname, 'fixtures', 'full-project')).toString(),
-        capabilities: {},
-        initializationOptions: {
-          isELSTesting: true,
-        },
-      };
+      const response = await initServer(connection);
 
-      const response = await connection.sendRequest((InitializeRequest.type as unknown) as string, params);
-
-      expect(response).toMatchSnapshot();
-    });
-  });
-
-  describe('Completion request', () => {
-    it('returns all components and helpers when requesting completion items in a handlebars expression', async () => {
-      const applicationTemplatePath = path.join(__dirname, 'fixtures', 'full-project', 'app', 'templates', 'application.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(applicationTemplatePath).toString(),
-        },
-        position: {
-          line: 1,
-          character: 2,
-        },
-      };
-
-      openFile(connection, applicationTemplatePath);
-
-      const response = await connection.sendRequest(CompletionRequest.type, params);
-
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns all angle-bracket in a element expression', async () => {
-      const applicationTemplatePath = path.join(__dirname, 'fixtures', 'full-project', 'app', 'templates', 'angle-completion.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(applicationTemplatePath).toString(),
-        },
-        position: {
-          line: 1,
-          character: 2,
-        },
-      };
-
-      openFile(connection, applicationTemplatePath);
-
-      const response = await connection.sendRequest(CompletionRequest.type, params);
-
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns all routes when requesting completion items in an inline link-to', async () => {
-      const templatePath = path.join(__dirname, 'fixtures', 'full-project', 'app', 'templates', 'definition.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(templatePath).toString(),
-        },
-        position: {
-          line: 2,
-          character: 23,
-        },
-      };
-
-      openFile(connection, templatePath);
-
-      const response = await connection.sendRequest(CompletionRequest.type, params);
-
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns all routes when requesting completion items in a block link-to', async () => {
-      const templatePath = path.join(__dirname, 'fixtures', 'full-project', 'app', 'templates', 'definition.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(templatePath).toString(),
-        },
-        position: {
-          line: 3,
-          character: 12,
-        },
-      };
-
-      openFile(connection, templatePath);
-
-      const response = await connection.sendRequest(CompletionRequest.type, params);
-
-      expect(response).toMatchSnapshot();
-    });
-  });
-
-  describe('Definition request', () => {
-    it('returns the definition information for a component in a template', async () => {
-      const base = path.join(__dirname, 'fixtures', 'full-project');
-      const definitionTemplatePath = path.join(base, 'app', 'templates', 'definition.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(definitionTemplatePath).toString(),
-        },
-        position: {
-          line: 0,
-          character: 4,
-        },
-      };
-
-      openFile(connection, definitionTemplatePath);
-
-      let response = await connection.sendRequest(DefinitionRequest.type, params);
-
-      response = normalizeUri(response, base);
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns the definition information for a helper in a template', async () => {
-      const base = path.join(__dirname, 'fixtures', 'full-project');
-      const definitionTemplatePath = path.join(base, 'app', 'templates', 'definition.hbs');
-      const params = {
-        textDocument: {
-          uri: URI.file(definitionTemplatePath).toString(),
-        },
-        position: {
-          line: 1,
-          character: 4,
-        },
-      };
-
-      openFile(connection, definitionTemplatePath);
-
-      let response = await connection.sendRequest(DefinitionRequest.type, params);
-
-      response = normalizeUri(response, base);
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns the definition information for a hasMany relationship', async () => {
-      const base = path.join(__dirname, 'fixtures', 'full-project');
-      const modelPath = path.join(base, 'app', 'models', 'model-a.js');
-      const params = {
-        textDocument: {
-          uri: URI.file(modelPath).toString(),
-        },
-        position: {
-          line: 4,
-          character: 27,
-        },
-      };
-
-      openFile(connection, modelPath);
-
-      let response = await connection.sendRequest(DefinitionRequest.type, params);
-
-      response = normalizeUri(response, base);
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns the definition information for a belongsTo relationship', async () => {
-      const base = path.join(__dirname, 'fixtures', 'full-project');
-      const modelPath = path.join(base, 'app', 'models', 'model-b.js');
-      const params = {
-        textDocument: {
-          uri: URI.file(modelPath).toString(),
-        },
-        position: {
-          line: 4,
-          character: 27,
-        },
-      };
-
-      openFile(connection, modelPath);
-
-      let response = await connection.sendRequest(DefinitionRequest.type, params);
-
-      response = normalizeUri(response, base);
-      expect(response).toMatchSnapshot();
-    });
-
-    it('returns the definition information for a transform', async () => {
-      const base = path.join(__dirname, 'fixtures', 'full-project');
-      const modelPath = path.join(base, 'app', 'models', 'model-a.js');
-      const params = {
-        textDocument: {
-          uri: URI.file(modelPath).toString(),
-        },
-        position: {
-          line: 6,
-          character: 27,
-        },
-      };
-
-      openFile(connection, modelPath);
-
-      let response = await connection.sendRequest(DefinitionRequest.type, params);
-
-      response = normalizeUri(response, base);
       expect(response).toMatchSnapshot();
     });
   });
